+++ godot-core/src/engine/mod.rs
@@ -19,10 +19,10 @@ pub use crate::gen::utilities;
 
 use crate::sys;
 
-mod gfile;
+mod io;
 mod script_instance;
 
-pub use gfile::{GFile, NotUniqueError};
+pub use io::*;
 pub use script_instance::{create_script_instance, ScriptInstance};
 
 #[cfg(debug_assertions)]
@@ -132,67 +132,6 @@ where
     }
 }
 
-/// Loads a resource from the filesystem located at `path`, panicking on error.
-///
-/// See [`try_load`] for more information.
-///
-/// # Example
-///
-/// ```no_run
-/// use godot::prelude::*;
-///
-/// let scene = load::<PackedScene>("res://path/to/Main.tscn");
-/// ```
-///
-/// # Panics
-/// If the resource cannot be loaded, or is not of type `T` or inherited.
-#[inline]
-pub fn load<T>(path: impl Into<GString>) -> Gd<T>
-where
-    T: GodotClass + Inherits<Resource>,
-{
-    let path = path.into();
-    load_impl(&path).unwrap_or_else(|| panic!("failed to load node at path `{path}`"))
-}
-
-/// Loads a resource from the filesystem located at `path`.
-///
-/// The resource is loaded on the method call (unless it's referenced already elsewhere, e.g. in another script or in the scene),
-/// which might cause slight delay, especially when loading scenes.
-///
-/// If the resource cannot be loaded, or is not of type `T` or inherited, this method returns `None`.
-///
-/// This method is a simplified version of [`ResourceLoader::load()`][crate::engine::ResourceLoader::load],
-/// which can be used for more advanced scenarios.
-///
-/// # Note:
-/// Resource paths can be obtained by right-clicking on a resource in the Godot editor (_FileSystem_ dock) and choosing "Copy Path",
-/// or by dragging the file from the _FileSystem_ dock into the script.
-///
-/// The path must be absolute (typically starting with `res://`), a local path will fail.
-///
-/// # Example
-/// Loads a scene called `Main` located in the `path/to` subdirectory of the Godot project and caches it in a variable.
-/// The resource is directly stored with type `PackedScene`.
-///
-/// ```no_run
-/// use godot::prelude::*;
-///
-/// if let Some(scene) = try_load::<PackedScene>("res://path/to/Main.tscn") {
-///     // all good
-/// } else {
-///     // handle error
-/// }
-/// ```
-// TODO Result to differentiate 2 errors
-#[inline]
-pub fn try_load<T>(path: impl Into<GString>) -> Option<Gd<T>>
-where
-    T: GodotClass + Inherits<Resource>,
-{
-    load_impl(&path.into())
-}
-
 // ----------------------------------------------------------------------------------------------------------------------------------------------
 // Utilities for crate
 
@@ -286,17 +225,4 @@ fn is_derived_base_cached(derived: ClassName, base: ClassName) -> bool {
     }
 
     is_parent_class
-}
-
-// Separate function, to avoid constructing string twice
-// Note that more optimizations than that likely make no sense, as loading is quite expensive
-fn load_impl<T>(path: &GString) -> Option<Gd<T>>
-where
-    T: GodotClass + Inherits<Resource>,
-{
-    ResourceLoader::singleton()
-        .load_ex(path.clone())
-        .type_hint(T::class_name().to_gstring())
-        .done() // TODO unclone
-        .and_then(|res| res.try_cast::<T>().ok())
 }
